+++ src/main/java/net/geforcemods/securitycraft/blocks/BlockChangeDetectorFloorCeilingBlock.java
@@ -24,7 +24,7 @@ public class BlockChangeDetectorFloorCeilingBlock extends BlockChangeDetectorBlo
 
 	@Override
 	public AxisAlignedBB getBoundingBox(IBlockState state, IBlockAccess world, BlockPos pos) {
-		IBlockState actualState = getDisguisedBlockState(world, pos);
+		IBlockState actualState = getDisguisedBlockState(world.getTileEntity(pos));
 
 		if (actualState != null && actualState.getBlock() != this)
 			return actualState.getBoundingBox(world, pos);