+++ src/main/java/net/geforcemods/securitycraft/blocks/BlockChangeDetectorWallBlock.java
@@ -32,7 +32,7 @@ public class BlockChangeDetectorWallBlock extends BlockChangeDetectorBlock {
 
 	@Override
 	public AxisAlignedBB getBoundingBox(IBlockState state, IBlockAccess world, BlockPos pos) {
-		IBlockState actualState = getDisguisedBlockState(world, pos);
+		IBlockState actualState = getDisguisedBlockState(world.getTileEntity(pos));
 
 		if (actualState != null && actualState.getBlock() != this)
 			return actualState.getBoundingBox(world, pos);