+++ src/main/java/net/geforcemods/securitycraft/blocks/DisguisableBlock.java
@@ -39,7 +39,7 @@ public class DisguisableBlock extends OwnableBlock implements IOverlayDisplay, I
 
 	@Override
 	public int getLightValue(IBlockState state, IBlockAccess world, BlockPos pos) {
-		IBlockState actualState = getDisguisedBlockState(world, pos);
+		IBlockState actualState = getDisguisedBlockState(world.getTileEntity(pos));
 
 		if (actualState != null && actualState.getBlock() != this)
 			return actualState.getLightValue(world, pos);
@@ -49,7 +49,7 @@ public class DisguisableBlock extends OwnableBlock implements IOverlayDisplay, I
 
 	@Override
 	public SoundType getSoundType(IBlockState state, World world, BlockPos pos, Entity entity) {
-		IBlockState actualState = getDisguisedBlockState(world, pos);
+		IBlockState actualState = getDisguisedBlockState(world.getTileEntity(pos));
 
 		if (actualState != null && actualState.getBlock() != this)
 			return actualState.getBlock().getSoundType(actualState, world, pos, entity);
@@ -59,7 +59,7 @@ public class DisguisableBlock extends OwnableBlock implements IOverlayDisplay, I
 
 	@Override
 	public AxisAlignedBB getBoundingBox(IBlockState state, IBlockAccess world, BlockPos pos) {
-		IBlockState actualState = getDisguisedBlockState(world, pos);
+		IBlockState actualState = getDisguisedBlockState(world.getTileEntity(pos));
 
 		if (actualState != null && actualState.getBlock() != this)
 			return actualState.getBoundingBox(world, pos);
@@ -69,7 +69,7 @@ public class DisguisableBlock extends OwnableBlock implements IOverlayDisplay, I
 
 	@Override
 	public AxisAlignedBB getCollisionBoundingBox(IBlockState state, IBlockAccess world, BlockPos pos) {
-		IBlockState actualState = getDisguisedBlockState(world, pos);
+		IBlockState actualState = getDisguisedBlockState(world.getTileEntity(pos));
 
 		if (actualState != null && actualState.getBlock() != this)
 			return actualState.getCollisionBoundingBox(world, pos);
@@ -79,7 +79,7 @@ public class DisguisableBlock extends OwnableBlock implements IOverlayDisplay, I
 
 	@Override
 	public AxisAlignedBB getSelectedBoundingBox(IBlockState state, World world, BlockPos pos) {
-		IBlockState actualState = getDisguisedBlockState(world, pos);
+		IBlockState actualState = getDisguisedBlockState(world.getTileEntity(pos));
 
 		if (actualState != null && actualState.getBlock() != this)
 			return actualState.getSelectedBoundingBox(world, pos);
@@ -89,7 +89,7 @@ public class DisguisableBlock extends OwnableBlock implements IOverlayDisplay, I
 
 	@Override
 	public void addCollisionBoxToList(IBlockState state, World world, BlockPos pos, AxisAlignedBB entityBox, List<AxisAlignedBB> collidingBoxes, Entity entity, boolean isActualState) {
-		IBlockState actualState = getDisguisedBlockState(world, pos);
+		IBlockState actualState = getDisguisedBlockState(world.getTileEntity(pos));
 
 		if (actualState != null && actualState.getBlock() != this)
 			actualState.addCollisionBoxToList(world, pos, entityBox, collidingBoxes, entity, true);
@@ -105,7 +105,7 @@ public class DisguisableBlock extends OwnableBlock implements IOverlayDisplay, I
 
 	@Override
 	public BlockFaceShape getBlockFaceShape(IBlockAccess world, IBlockState state, BlockPos pos, EnumFacing face) {
-		return IDisguisable.getDisguisedBlockFaceShape(world, state, pos, face);
+		return IDisguisable.getDisguisedBlockFaceShape(world, pos, face);
 	}
 
 	@Override
@@ -133,7 +133,7 @@ public class DisguisableBlock extends OwnableBlock implements IOverlayDisplay, I
 
 	@Override
 	public IBlockState getActualState(IBlockState state, IBlockAccess world, BlockPos pos) {
-		IBlockState disguisedState = getDisguisedBlockState(world, pos);
+		IBlockState disguisedState = getDisguisedBlockState(world.getTileEntity(pos));
 
 		return disguisedState != null ? disguisedState : state;
 	}