+++ src/main/java/net/geforcemods/securitycraft/blocks/KeypadChestBlock.java
@@ -77,7 +77,7 @@ public class KeypadChestBlock extends DisguisableBlock {
 
 	@Override
 	public AxisAlignedBB getBoundingBox(IBlockState state, IBlockAccess world, BlockPos pos) {
-		IBlockState actualState = getDisguisedBlockState(world, pos);
+		IBlockState actualState = getDisguisedBlockState(world.getTileEntity(pos));
 
 		if (actualState != null && actualState.getBlock() != this)
 			return actualState.getBoundingBox(world, pos);