+++ src/main/java/net/geforcemods/securitycraft/blocks/reinforced/ReinforcedDispenserBlock.java
@@ -104,7 +104,7 @@ public class ReinforcedDispenserBlock extends BlockDispenser implements IReinfor
 
 	@Override
 	public int getLightValue(IBlockState state, IBlockAccess world, BlockPos pos) {
-		IBlockState actualState = getDisguisedBlockState(world, pos);
+		IBlockState actualState = getDisguisedBlockState(world.getTileEntity(pos));
 
 		if (actualState != null && actualState.getBlock() != this)
 			return actualState.getLightValue(world, pos);
@@ -114,7 +114,7 @@ public class ReinforcedDispenserBlock extends BlockDispenser implements IReinfor
 
 	@Override
 	public SoundType getSoundType(IBlockState state, World world, BlockPos pos, Entity entity) {
-		IBlockState actualState = getDisguisedBlockState(world, pos);
+		IBlockState actualState = getDisguisedBlockState(world.getTileEntity(pos));
 
 		if (actualState != null && actualState.getBlock() != this)
 			return actualState.getBlock().getSoundType(actualState, world, pos, entity);
@@ -124,7 +124,7 @@ public class ReinforcedDispenserBlock extends BlockDispenser implements IReinfor
 
 	@Override
 	public AxisAlignedBB getBoundingBox(IBlockState state, IBlockAccess world, BlockPos pos) {
-		IBlockState actualState = getDisguisedBlockState(world, pos);
+		IBlockState actualState = getDisguisedBlockState(world.getTileEntity(pos));
 
 		if (actualState != null && actualState.getBlock() != this)
 			return actualState.getBoundingBox(world, pos);
@@ -134,7 +134,7 @@ public class ReinforcedDispenserBlock extends BlockDispenser implements IReinfor
 
 	@Override
 	public void addCollisionBoxToList(IBlockState state, World world, BlockPos pos, AxisAlignedBB entityBox, List<AxisAlignedBB> collidingBoxes, Entity entity, boolean isActualState) {
-		IBlockState actualState = getDisguisedBlockState(world, pos);
+		IBlockState actualState = getDisguisedBlockState(world.getTileEntity(pos));
 
 		if (actualState != null && actualState.getBlock() != this)
 			actualState.addCollisionBoxToList(world, pos, entityBox, collidingBoxes, entity, true);
@@ -144,7 +144,7 @@ public class ReinforcedDispenserBlock extends BlockDispenser implements IReinfor
 
 	@Override
 	public BlockFaceShape getBlockFaceShape(IBlockAccess world, IBlockState state, BlockPos pos, EnumFacing face) {
-		return IDisguisable.getDisguisedBlockFaceShape(world, state, pos, face);
+		return IDisguisable.getDisguisedBlockFaceShape(world, pos, face);
 	}
 
 	@Override
@@ -155,7 +155,7 @@ public class ReinforcedDispenserBlock extends BlockDispenser implements IReinfor
 
 	@Override
 	public IBlockState getActualState(IBlockState state, IBlockAccess world, BlockPos pos) {
-		IBlockState disguisedState = getDisguisedBlockState(world, pos);
+		IBlockState disguisedState = getDisguisedBlockState(world.getTileEntity(pos));
 
 		return disguisedState != null ? disguisedState : state;
 	}