+++ Changelog.md
@@ -1,6 +1,7 @@
 --------------------------Changelog for the upcoming version of SecurityCraft--------------------------
 
 - Change: The block mine overlay now shows in the first person hand models as well
+- API: IDisguisable has been simplified
 
 --------------------------Changelog for v1.9.12 of SecurityCraft--------------------------